+++ src/modules/esm.mjs
@@ -1,0 +1,40 @@
+import path from 'node:path';
+import { release, version } from 'node:os';
+import { createServer as createServerHttp } from 'node:http';
+import './files/c.js';
+import { readFile } from 'node:fs/promises';
+import { fileURLToPath } from 'node:url';
+
+const random = Math.random();
+
+let unknownObject;
+
+const aJson = JSON.parse(
+  await readFile(new URL('./files/a.json', import.meta.url))
+);
+const bJson = JSON.parse(
+  await readFile(new URL('./files/b.json', import.meta.url))
+);
+unknownObject = random > 0.5 ? aJson : bJson;
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log('To terminate it, use Ctrl+C combination');
+});
